Added Tour and Enumconverter
Additions:
- CreateTour
- ETransportTypeConverter

+++ SWE_TourPlanner_WPF/SWE_TourPlanner_WPF/Tour.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Documents;
@@ -14,15 +16,15 @@ namespace SWE_TourPlanner_WPF
         public string From { get; set; }
         public string To { get; set; }
         public ETransportType TransportType { get; set; }
-        public float Distance { get; set; }
-        public float Time { get; set; }
+        public int Distance { get; set; }
+        public int Time { get; set; }
         public string RouteInformation { get; set; }
         public List<TourLog> TourLogs { get; set; } = new List<TourLog>();
 
 
         public Tour() { }
 
-        public Tour(string name, string description, string from, string to, ETransportType transportType, float distance, float time, string routeInformation, List<TourLog> tourLogs)
+        public Tour(string name, string description, string from, string to, ETransportType transportType, int distance, int time, string routeInformation, List<TourLog> tourLogs)
         {
             Name = name;
             Description = description;
